backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
langchain==0.3.25
langchain-anthropic==0.3.12
langgraph==0.4.7
crewai==0.120.1
autogen-agentchat==0.4.12
mcp==1.9.3
httpx==0.28.1
